+++ docs/changelog.md
@@ -16,6 +16,7 @@
 + Move `icons` folder into `pyslvs_ui/icons`.
 + Replace `*-pyslvs` command with `*-kernel` in Makefile.
 + Treat `python-solvespace` as normal module.
++ Refactor `main_canvas` to `canvas`.
 
 ## Version 19.09